Confidence Intervals for Model Parameters

Computes confidence intervals for one or more parameters in a fitted model. This is a method specific to the "gam" class from package "mgcv".

# S3 method for gam
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object
a fitted model object of class "gam".

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
the confidence level required.

...
not implemented

Value

A tidy data frame containing parameter names, Estimates and confidence intervals for parametric terms

Examples

set.seed(101)
library(dplyr)
library(mgcv)
dat <- data.frame(x = runif(n=100),
                  y = runif(n=100)) %>%
  dplyr::mutate(z = rnorm(n=100,
                   mean = 1 - 2*x - sin(2*pi*y),
                   sd = 0.1))

fit1 <- gam(data=dat, z ~ y + s(x))

confint(fit1)

#> Joining, by = "term"

#> # A tibble: 2 x 5
#>          term   Estimate `Std. Error`    `2.5%`    `97.5%`
#>         <chr>      <dbl>        <dbl>     <dbl>      <dbl>
#> 1 (Intercept) -0.9494979   0.09135369 -1.130810 -0.7681861
#> 2           y  1.7176761   0.15180645  1.416382  2.0189699

confint(fit1, parm="y", level=0.8)

#> Joining, by = "term"

#> # A tibble: 1 x 5
#>    term Estimate `Std. Error`  `10.0%`  `90.0%`
#>   <chr>    <dbl>        <dbl>    <dbl>    <dbl>
#> 1     y 1.717676    0.1518064 1.521794 1.913558