Tidy Data Frame of Smooth Terms

Returns a tidy data frame containing the mean and confidence intervals for fitted smooths from a model of class "gam" from package "mgcv".

tidy_smooths(object, dimension = 1, level = 0.95, parm = NULL)

Arguments

object
a fitted model object of class "gam".

dimension
the dimension of the smooths that are desired for extraction.

level
the confidence level required.

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered (not yet implemented).

...
not implemented

Value

A tidy data frame containing parameter names, estimates and confidence intervals for non-parametric terms

Examples

set.seed(101)
library(dplyr)
library(mgcv)
library(mgcv.helper)
dat <- data.frame(x = runif(n=100),
                  y = runif(n=100)) %>%
  dplyr::mutate(z = rnorm(n=100,
                   mean = 1 - 2*x - sin(2*pi*y),
                   sd = 0.1))

fit1 <- gam(data=dat, z ~ y + s(x))

tidy_smooths(fit1, dimension=1)

#> # A tibble: 100 x 6
#>             x         y      ymin     ymax   xlab   ylab
#>         <dbl>     <dbl>     <dbl>    <dbl> <fctr> <fctr>
#>  1 0.01018864 0.9574051 0.6449827 1.269828      x s(x,1)
#>  2 0.02012017 0.9389361 0.6325405 1.245332      x s(x,1)
#>  3 0.03005170 0.9204671 0.6200983 1.220836      x s(x,1)
#>  4 0.03998322 0.9019981 0.6076562 1.196340      x s(x,1)
#>  5 0.04991475 0.8835290 0.5952140 1.171844      x s(x,1)
#>  6 0.05984627 0.8650600 0.5827718 1.147348      x s(x,1)
#>  7 0.06977780 0.8465910 0.5703296 1.122852      x s(x,1)
#>  8 0.07970932 0.8281220 0.5578875 1.098356      x s(x,1)
#>  9 0.08964085 0.8096529 0.5454453 1.073861      x s(x,1)
#> 10 0.09957237 0.7911839 0.5330031 1.049365      x s(x,1)
#> # ... with 90 more rows