Variance Inflation Factor

This function takes a fitted mgcv model object and returns a data frame of variance inflation factors

vif.gam(object)

Arguments

object
An object of class gam

Value

VIF.df A data frame consisting of the VIF values for each parametric term in a fitted Generalised Additive Model

Examples

library(mgcv)
library(dplyr)

set.seed(101)
N <- 100
x1 <- runif(n=N)
x2 <- runif(n=N)
x3 <- runif(n=N) + 0.9*x1 - 1.75*x2

df <- data.frame(x1 = x1,
                 x2 = x2,
                 x3 = x3) %>%
  mutate(y = rnorm(n=N,
                   mean = 1 - 2*x1 + 3*x2 - 0.5*x3,
                   sd = 0.5))

fit1 <- gam(data=df, y ~ x1 + x2 + x3)

summary(fit1)

#> 
#> Family: gaussian 
#> Link function: identity 
#> 
#> Formula:
#> y ~ x1 + x2 + x3
#> 
#> Parametric coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   0.9966     0.1410   7.071 2.48e-10 ***
#> x1           -1.7966     0.2215  -8.111 1.65e-12 ***
#> x2            2.8581     0.3377   8.464 2.94e-13 ***
#> x3           -0.6007     0.1652  -3.637 0.000446 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> 
#> R-sq.(adj) =  0.891   Deviance explained = 89.5%
#> GCV = 0.22009  Scale est. = 0.21129   n = 100

vif.gam(fit1)

#> # A tibble: 3 x 2
#>   variable      vif
#>      <chr>    <dbl>
#> 1       x1 2.047334
#> 2       x2 4.900343
#> 3       x3 5.393642